backend/pyproject.toml
[tool.poetry]
name = "eco-sight"
version = "0.1.0"
description = "Environmental information management application"
authors = ["Samara Marques <samarassk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
requests = "^2.32.5"
reportlab = "^4.4.4"
python-dotenv = "^1.2.1"
djangorestframework = "^3.16.1"
matplotlib = "^3.10.7"
pytz = "^2025.2"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
